requirements.txt
Django
djangorestframework
django-crispy-forms
django-filter
